requirements.txt
# Core ML and NLP libraries
transformers==4.40.0
datasets==2.18.0
torch==2.7.1+cu118
numpy==1.26.4
scikit-learn==1.5.2

# Explainability libraries
lime==0.2.0.1
shap==0.45.0

# Web app and utilities
streamlit==1.36.0
pandas==2.3.0
tqdm==4.67.1

# Configuration and logging
pyyaml==6.0.2
click==8.1.7

# Text processing
nltk==3.8.1
spacy==3.8.0

# Development and testing
pytest==8.2.0
flake8==7.0.0
black==24.10.0

# Additional dependencies
﻿accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudpathlib==0.21.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
coverage==7.6.8
customtkinter==5.2.2
cycler==0.12.1
cymem==2.0.11
darkdetect==0.8.0
dateparser==1.2.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.18.0
fsspec==2024.2.0
huggingface-hub==0.33.4
idna==3.10
importlib-metadata==7.1.0
importlib-resources==6.4.0
ipykernel==6.29.5
ipython==8.25.0
ipython-genutils==0.2.0
jedi==0.19.1
jinja2==3.1.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.02
jupyter-client==9.0.1
jupyter-core==5.7.2
jupyterlab-pygments==0.3.0
kiwisolver==1.4.5
markupsafe==2.1.5
matplotlib==3.9.1
mistune==3.0.2
mpmath==1.4.0
multidict==6.4.4
multiprocess==0.70.16
nbclient==0.10.0
nbconvert==8.2.0
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.1.3
notebook-shim==0.2.4
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
platformdirs==4.3.0
prometheus-client==0.20.0
prompt-toolkit==3.0.43
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.22
pygments==2.18.0
pyparsing==3.1.2
pyrsistent==0.21.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==26.0.2
qtconsole==5.5.1
qtpy==2.4.1
referencing==0.34.0
requests==2.32.4
rpds-py==0.18.1
send2trash==1.8.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.13
terminado==0.18.0
tinycss2==1.3.0
tokenizers==0.20.3
toml==0.10.2
tornado==6.4
traitlets==5.14.2
typing-extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.2
wcwidth==0.2.13
webencodings==0.5.1
xxhash==3.5.0
yarl==1.20.0
zipp==3.19.2